Slight improvements to display format (handle narrow displays, avoid wraaround in possible 32-bit generation counter)

+++ trunk/SpeedComparisons/GrayScott_HWIVector/display_hwiv.h
@@ -1,3 +1,3 @@
 bool display(int width, int height, float *r, float *g, float *b,
-             int iteration, float model_scale, bool auto_brighten,float manual_brighten,
+             double iteration, float model_scale, bool auto_brighten,float manual_brighten,
 			 int scale,int delay_ms,const char* message, bool write_video);
